requirements.txt
torch
torchvision
PyQt6
opencv-python
pillow
numpy
matplotlib 